+++ src/App.tsx
@@ -2,10 +2,11 @@ import './App.scss';
 
 import React from 'react';
 import Home from './pages/home/home.component';
-import FearGreedIndex from './components/fear-greed-index/fear-greed-index.component';
 import Navbar from './components/navbar/navbar.component';
 import Topbar from './components/topbar/topbar.component';
 import Market from './pages/market/market.component';
+import Learn from './pages/learn/learn.component';
+import FearGreedIndex from './components/fear-greed-index/fear-greed-index.component';
 
 function App() {
 	return (
@@ -14,6 +15,7 @@ function App() {
 			<Navbar />
 			<Home />
 			<Market />
+			<Learn />
 			<FearGreedIndex />
 		</div>
 	);