+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 - Added pagination to 'Market' table (displaying 10 coins per page)
 - Added ability to sort market table columns
 - Added ability to select the number of rows displayed per page via dropdown
+- Added initial 'Learn' section